refactor: move display folder into the cli module
