+++ src/mini_agent/cli/display/__init__.py
