+++ src/mini_agent/cli/display/completion.py
