+++ src/mini_agent/cli/display/diff.py
